Modified Makefile to use non-executable stack.
Modified the link rule so that it explicitly states that the stack
should be non-executable.  The newer toolchain on the CSUG machines
generates a warning because the assembly code is missing a directive
indicating that the code will work with a non-executable stack, and
adding this directive prevents assembly.  Telling the linker what
to do fixes the issue.

Switched the compiler from GCC to Clang.